+++ contracts/mocks/token/ERC20/ERC677Mock.sol
@@ -5,7 +5,7 @@ import {IForwarderRegistry} from "./../../../metatx/interfaces/IForwarderRegistr
 import {IERC677} from "./../../../token/ERC20/interfaces/IERC677.sol";
 import {IERC677Receiver} from "./../../../token/ERC20/interfaces/IERC677Receiver.sol";
 import {ERC20Storage} from "./../../../token/ERC20/libraries/ERC20Storage.sol";
-import {Address} from "@openzeppelin/contracts/utils/Address.sol";
+import {Address} from "./../../../utils/libraries/Address.sol";
 import {ERC20FixedSupply} from "./../../../token/ERC20/preset/ERC20FixedSupply.sol";
 
 contract ERC677Mock is ERC20FixedSupply, IERC677 {
@@ -24,7 +24,7 @@ contract ERC677Mock is ERC20FixedSupply, IERC677 {
     function transferAndCall(address receiver, uint256 amount, bytes calldata data) external returns (bool success) {
         address sender = _msgSender();
         ERC20Storage.layout().transfer(sender, receiver, amount);
-        if (!receiver.isContract()) return true;
+        if (!receiver.hasBytecode()) return true;
         return IERC677Receiver(receiver).onTokenTransfer(sender, amount, data);
     }
 }