+++ test/contracts/security/TokenRecovery.test.js
@@ -66,7 +66,9 @@ runBehaviorTests('TokenRecovery', config, function (deployFn) {
     });
 
     it('reverts with an amount above the contract balance', async function () {
-      await expect(this.contract.recoverETH([deployer.address], ['1001'])).to.be.revertedWith('Address: insufficient balance');
+      await expect(this.contract.recoverETH([deployer.address], ['1001']))
+        .to.be.revertedWithCustomError(this.contract, 'AddressInsufficientBalance')
+        .withArgs(await this.contract.getAddress());
     });
 
     context('when successful', function () {